recommended to add examples of usermetadata
The scheme of saving the mapping relationship through sql/nosql external storage has advantages and disadvantages, the main disadvantage is that it introduces a heavy dependency and the complexity is relatively high, so many people consider adding information directly to the minio file "usermetadata". Listing examples of object information containing "usermetadata" is quite important, the official document or examples does not introduce it. It is easy to cause the user to use "listObjects" storage file and do not know that this attribute needs to be enabled. As a result,"io.minio.messages.Item" does not have "usermetadata"